Xilsecure: Literal value requires a U suffix.
This patch fixes below MISRA-C violation:
Literal value requires a U suffix.
MISRA-C:2012/AMD1 R.7.2,R.10.3,R.10.4


Acked-for-series: VNSL Durga Challa <vnsl.durga.challa@xilinx.com>

+++ lib/sw_services/xilsecure/src/zynqmp/xsecure_rsa_core.c
@@ -87,7 +87,7 @@ u32 XSecure_RsaCfgInitialize(XSecure_Rsa *InstancePtr)
 	u32 Status;
 
 	InstancePtr->BaseAddress = XSECURE_CSU_RSA_BASE;
-	Status = XST_SUCCESS;
+	Status = (u32)XST_SUCCESS;
 
 	return Status;
 }
@@ -133,7 +133,7 @@ u32 XSecure_RsaOperation(XSecure_Rsa *InstancePtr, u8 *Input,
 			(Size == XSECURE_RSA_4096_KEY_SIZE));
 
 	InstancePtr->EncDec = EncDecFlag;
-	InstancePtr->SizeInWords = Size/4;
+	InstancePtr->SizeInWords = Size/4U;
 	/* Put Modulus, exponent, Mod extension in RSA RAM */
 	XSecure_RsaPutData(InstancePtr);
